requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
asgiref==3.2.10
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.1.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Django==3.1
Flask==1.1.2
gast==0.2.2
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
image==1.5.32
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
multidict==4.7.6
numpy==1.19.1
oauthlib==3.1.0
opencv-python==4.4.0.42
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
sqlparse==0.3.1
tensorboard==1.15.0
tensorboard-plugin-wit==1.7.0
tensorflow==1.15.3
tensorflow-estimator==1.15.1
termcolor==1.1.0
typing-extensions==3.7.4.2
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.5.1
zipp==3.1.0
